High-level API

The density function $\ell$: an example

The density function should take a single argument q, which is a vector of numbers, and return an object which provides the methods DiffResults.value and DiffResults.gradient to access $\ell(q)$ and $\nabla\ell(q)$, respectively.

The following example implements the density function for $n$ observations from a $\text{Bernoulli}(\alpha)$ distribution, $s$ of which are 1. This example and others are available in DynamicHMCExamples.jl.

First, we load some packages we use (you may have to install them):

using DynamicHMC
using ContinuousTransformations
using Parameters
using MCMCDiagnostics
using DiffWrappers

It is convenient to define a structure that holds the data,

struct BernoulliProblem
    "Total number of draws in the data."
    n::Int
    "Number of draws =1 in the data."
    s::Int
end

then make it callable (with a vector argument):

function (problem::BernoulliProblem)(θ)
    α, = θ                        # extract the first parameter
    @unpack n, s = problem        # using Parameters
    s * log(α) + (n-s) * log(1-α) # log likelihood
end

and finally define an object with actual data:

p = BernoulliProblem(100, 40)                             # original problem

The value p is a function that takes a single real number, and returns the likelihood. However, the functions in this package

take a vector which contains elements in $\mathbb{R}$, and

expect $\ell$ (ie p above) to return an object that can provide the value and the derivatives.

We could implement both manually, but it is convenient to use wrappers from two packages mentioned in the introduction:

pt = TransformLogLikelihood(p, bridge(ℝ, Segment(0, 1)))  # transform
pt∇ = ForwardGradientWrapper(pt, [0.0]);                  # AD using ForwardDiff.jl

Then we can call the high-level function NUTS_init_tune_mcmc that initializes and tunes the sampler, and samples from it:

sample, NUTS_tuned = NUTS_init_tune_mcmc(pt∇, 1, 1000);

The returned objects are the sample, which contains the draws and diagnostic information, and the tuned sampler, which we could use to continue sampling.

We obtain the posterior using the transformation and get_position:

posterior = map(get_transformation(pt) ∘ get_position, sample);

which is a vector of vectors. Calculate the effective sample size and NUTS-specific statistics as

julia> effective_sample_size(first.(posterior))
323.6134099739428

julia> NUTS_statistics(sample)         # NUTS-specific statistics
Hamiltonian Monte Carlo sample of length 1000
  acceptance rate mean: 0.92, min/25%/median/75%/max: 0.26 0.87 0.97 1.0 1.0
  termination: AdjacentTurn => 31% DoubledTurn => 69%
  depth: 1 => 66% 2 => 34%

Sampling and accessors

Most users would use this function, which initializes and tunes the parameters of the algorithm, then samples. Parameters can be set manually for difficult posteriors.

DynamicHMC.NUTS_init_tune_mcmc — Function.

NUTS_init_tune_mcmc(rng, ℓ, q_or_dim, N; args...)

Init, tune, and then draw N samples from ℓ using the NUTS algorithm.

Return the sample (a vector of NUTS_transitions) and the tuned sampler.

rng is the random number generator.

q_or_dim is a starting position or the dimension (for random initialization).

args are passed on to various methods, see NUTS_init and bracketed_doubling_tuner.

For parameters q, ℓ(q) should return an object that support the following methods: DiffResults.value, DiffResults.gradient.

Most users would use this function, unless they are doing something that requires manual tuning.

source

NUTS_init_tune_mcmc(ℓ, q_or_dim, N; args...)

Same as the other method, but with random number generator Base.Random.GLOBAL_RNG.

source

Important

The NUTS sampler saves a random number generator and uses it for random draws. When running in parallel, you should initialize NUTS_init_tune_mcmc with a random number generator as its first argument explicitly, making sure that each thread has its own one.

These functions can be used use to perform the steps above manually.

DynamicHMC.NUTS_init — Function.

NUTS_init(rng, ℓ, q; κ = GaussianKE(length(q)), p, max_depth, ϵ)

Initialize a NUTS sampler for log density ℓ using local information.

Arguments

rng: the random number generator

ℓ: the likelihood function, should return a type that supports DiffResults.value and DiffResults.gradient

q: initial position.

κ: kinetic energy specification. Default: Gaussian with identity matrix.

p: initial momentum. Default: random from standard multivariate normal.

max_depth: maximum tree depth. Default: 5.

ϵ: initial stepsize, or parameters for finding it (passed on to find_initial_stepsize.

source

NUTS_init(rng, ℓ, dim::Integer; args...)

Random initialization with position randn(dim), all other arguments are passed on the the other method of this function.

source

DynamicHMC.tune — Function.

sampler′ = tune(sampler, tune)

Given a sampler (or similar a parametrization) and a tuner, return the updated sampler state after tuning.

source

DynamicHMC.mcmc — Function.

mcmc(sampler, N)

Run the MCMC sampler for N iterations, returning the results as a vector, which has elements that conform to the sampler.

source

The resulting sample is a vector of NUTS_Transition objects, for which the following accessors exist:

DynamicHMC.NUTS_Transition — Type.

Single transition by the No-U-turn sampler. Contains new position and diagnostic information.

source

DynamicHMC.get_position — Function.

Position after transition.

source

DynamicHMC.get_neg_energy — Function.

Negative energy of the Hamiltonian at the position.

source

DynamicHMC.get_depth — Function.

Tree depth.

source

DynamicHMC.get_termination — Function.

Reason for termination, see Termination.

source

DynamicHMC.get_acceptance_rate — Function.

get_acceptance_rate(x)

Return average Metropolis acceptance rate.

source

Average acceptance rate over trajectory.

source

DynamicHMC.get_steps — Function.

Number of integrator steps.

source

DynamicHMC.get_position_matrix — Function.

variable_matrix(posterior)

Return the samples of the parameter vector as rows of a matrix.

source

Diagnostics

These are NUTS-specific diagnostics and statistics (except for sample_cov, which is a convenience function). It is also prudent to use generic MCMC convergence diagnostics, as suggested in the introduction.

DynamicHMC.NUTS_statistics — Function.

NUTS_statistics(sample)

Return statistics about the sample (ie not the variables). Mostly useful for NUTS diagnostics.

source

DynamicHMC.sample_cov — Function.

sample_cov(sample)

Covariance matrix of the sample.

source

DynamicHMC.EBFMI — Function.

EBFMI(sample)

Energy Bayesian fraction of missing information. Useful for diagnosing poorly chosen kinetic energies.

Low values (≤ 0.3) are considered problematic. See Betancourt (2016).

source

Fine-grained control

Kinetic energy

DynamicHMC.KineticEnergy — Type.

Kinetic energy specifications.

For all subtypes, it is assumed that kinetic energy is symmetric in the momentum p, ie.

neg_energy(::KineticEnergy, p, q) == neg_energy(::KineticEnergy, -p, q)

When the above is violated, various implicit assumptions will not hold.

source

DynamicHMC.EuclideanKE — Type.

Euclidean kinetic energies (position independent).

source

DynamicHMC.GaussianKE — Type.

Gaussian kinetic energy.

\[p \mid q ∼ \text{Normal}(0, M) \qquad (\text{importantly, independent of $q$})\]

The inverse covariance M⁻¹ is stored.

source

NUTS parameters and tuning

DynamicHMC.NUTS — Type.

Specification for the No-U-turn algorithm, including the random number generator, Hamiltonian, the initial position, and various parameters.

source

DynamicHMC.AbstractTuner — Type.

A tuner that adapts the sampler.

All subtypes support length which returns the number of steps (note: if not in field N, define length accordingly), other parameters vary.

source

DynamicHMC.StepsizeTuner — Type.

Adapt the integrator stepsize for N samples.

source

DynamicHMC.StepsizeCovTuner — Type.

Tune the integrator stepsize and covariance. Covariance tuning is from scratch (no prior information is used), regularized towards the identity matrix.

source

DynamicHMC.TunerSequence — Type.

Sequence of tuners, applied in the given order.

source

DynamicHMC.mcmc_adapting_ϵ — Function.

sample, A = mcmc_adapting_ϵ(rng, sampler, N, [A_params, A])

Same as mcmc, but tune stepsize ϵ according to the parameters A_params and initial state A. Return the updated A as the second value.

When the last two parameters are not specified, initialize using adapting_ϵ.

source

DynamicHMC.bracketed_doubling_tuner — Function.

bracketed_doubling_tuner(; [init], [mid], [M], [term], [regularize])

A sequence of tuners:

tuning stepsize with init steps

tuning stepsize and covariance: first with mid steps, then repeat with twice the steps M times

tuning stepsize with term steps

regularize is used for covariance regularization.

source