Low-level building blocks

Notation

Notation follows Betancourt (2017), with some differences.

Instead of energies, negative energies are used in the code.

The following are used consistently for variables:

ℓ: log density we sample from, see this explanation

κ: distribution/density that corresponds to kinetic energy

H: Hamiltonian

q: position

p: momentum

z: point in phase space (q,p)

ϵ: stepsize

a: acceptance rate

A: acceptance tuning state

ζ: proposal from trajectory (phase point and weight)

τ: turn statistic

d: divergence statistic

π: log density (different from papers)

Δ: logdensity relative to initial point of trajectory

Low-level building blocks

This is documented only for package developers. These are not part of the public API, if you are using them you should reconsider or file an issue.

Hamiltonian and leapfrog

DynamicHMC.Hamiltonian — Type.

Hamiltonian(ℓ, κ)

Construct a Hamiltonian from the log density ℓ, and the kinetic energy specification κ. Calls of ℓ with a vector are expected to return a value that supports DiffResults.value and DiffResults.gradient.

source

DynamicHMC.PhasePoint — Type.

A point in phase space, consists of a position and a momentum.

Log densities and gradients are saved for speed gains, so that the gradient of ℓ at q is not calculated twice for every leapfrog step (both as start- and endpoints).

Because of caching, a PhasePoint should only be used with a specific Hamiltonian.

source

DynamicHMC.get_ℓq — Function.

get_ℓq(z)

The value returned by ℓ when evaluated at position q.

source

DynamicHMC.phasepoint_in — Function.

phasepoint_in(H::Hamiltonian, q, p)

The recommended interface for creating a phase point in a Hamiltonian. Computes cached values.

source

DynamicHMC.rand_phasepoint — Function.

rand_phasepoint(rng, H, q)

Extend a position q to a phasepoint with a random momentum according to the kinetic energy of H.

source

DynamicHMC.neg_energy — Function.

neg_energy(κ, p, [q])

Return the log density of kinetic energy κ, at momentum p. Some kinetic energies (eg Riemannian geometry) will need q, too.

source

neg_energy(H, z)

Log density for Hamiltonian H at point z.

If ℓ(q) == -Inf (rejected), ignores the kinetic energy.

source

DynamicHMC.get_p♯ — Function.

get_p♯(κ, p, [q])

Return $psharp$, used for turn diagnostics.

source

DynamicHMC.leapfrog — Function.

leapfrog(H, z, ϵ)

Take a leapfrog step of length ϵ from z along the Hamiltonian H.

Return the new position.

The leapfrog algorithm uses the gradient of the next position to evolve the momentum. If this is not finite, the momentum won't be either. Since the constructor PhasePoint validates its arguments, this can only happen for divergent points anyway, and should not cause a problem.

source

DynamicHMC.is_valid_ℓq — Function.

is_valid_ℓq(ℓq)

Test that a value returned by ℓ is valid, in the following sense:

supports DiffResults.value and DiffResults.gradient (when not, a

MethodError is thrown),

the value is a float, either -Inf or finite,

the gradient is finite when the value is; otherwise the gradient is ignored.

source

Finding initial stepsize $\epsilon$

Local stepsize tuning.

The local acceptance ratio is technically a probability, but for finding the initial stepsize, it is not capped at $1$.

Also, the values are cached as this is assumed to be moderately expensive to calculate.

DynamicHMC.find_initial_stepsize — Function.

find_initial_stepsize(parameters, A)

Find an initial stepsize that matches the conditions of parameters (see InitialStepsizeSearch).

A is the local acceptance ratio (uncapped). When given a Hamiltonian H and a phasepoint z, it will be calculated using local_acceptance_ratio.

source

DynamicHMC.InitialStepsizeSearch — Type.

Parameters for the search algorithm for the initial stepsize.

The algorithm finds an initial stepsize $ϵ$ so that the local acceptance ratio $A(ϵ)$ satisfies

\[a_\text{min} ≤ A(ϵ) ≤ a_\text{max}\]

This is achieved by an initial bracketing, then bisection.

a_min

Lowest local acceptance rate.

a_max

Highest local acceptance rate.

ϵ₀

Initial stepsize.

C

Scale factor for initial bracketing, > 1. Default: 2.0.

maxiter_crossing

Maximum number of iterations for initial bracketing.

maxiter_bisect

Maximum number of iterations for bisection.

Note

Cf. Hoffman and Gelman (2014), which does not ensure bounds for the acceptance ratio, just that it has crossed a threshold. This version seems to work better for some tricky posteriors with high curvature.

source

DynamicHMC.find_crossing_stepsize — Function.

Find the stepsize for which the local acceptance rate A(ϵ) crosses a.

find_crossing_stepsize(parameters, A, ϵ₀)
find_crossing_stepsize(parameters, A, ϵ₀, Aϵ₀)

Return ϵ₀, A(ϵ₀), ϵ₁, A(ϵ₁), whereϵ₀andϵ₁are stepsizes before and after crossingawithA(ϵ)`, respectively.

Assumes that $A(ϵ₀) ∉ (a_\text{min}, a_\text{max})$, where the latter are defined in parameters.

parameters: parameters for the iteration.

A: local acceptance ratio (uncapped), a function of stepsize ϵ

ϵ₀, Aϵ₀: initial value of ϵ, and A(ϵ₀)

source

DynamicHMC.bisect_stepsize — Function.

Return the desired stepsize ϵ by bisection.

bisect_stepsize(parameters, A, ϵ₀, ϵ₁)
bisect_stepsize(parameters, A, ϵ₀, ϵ₁, Aϵ₀)
bisect_stepsize(parameters, A, ϵ₀, ϵ₁, Aϵ₀, Aϵ₁)

parameters: algorithm parameters, see InitialStepsizeSearch

A: local acceptance ratio (uncapped), a function of stepsize ϵ

ϵ₀, ϵ₁, Aϵ₀, Aϵ₁: stepsizes and acceptance rates (latter optional).

This function assumes that $ϵ₀ < ϵ₁$, the stepsize is not yet acceptable, and the cached A values have the correct ordering.

source

DynamicHMC.local_acceptance_ratio — Function.

local_acceptance_ratio(H, z)

Return a function of the stepsize ($ϵ$) that calculates the local acceptance ratio for a single leapfrog step around z along the Hamiltonian H. Formally, let

\[A(ϵ) = \exp(\text{neg_energy}(H, \text{leapfrog}(H, z, ϵ)) - \text{neg_energy}(H, z))\]

Note that the ratio is not capped by 1, so it is not a valid probability per se.

source

Dual averaging

DynamicHMC.DualAveragingParameters — Type.

Parameters for the dual averaging algorithm of Gelman and Hoffman (2014, Algorithm 6).

To get reasonable defaults, initialize with DualAveragingParameters(logϵ₀). See adapting_ϵ for a joint constructor.

source

DynamicHMC.DualAveragingAdaptation — Type.

Current state of adaptation for ϵ. Use DualAverageingAdaptation(logϵ₀) to get an initial value. See adapting_ϵ for a joint constructor.

source

DynamicHMC.get_ϵ — Function.

get_ϵ(A, tuning = true)

When tuning, return the stepsize ϵ for the next HMC step. Otherwise return the tuned ϵ.

source

DynamicHMC.adapting_ϵ — Function.

DA_params, A = adapting_ϵ(ϵ; args...)

Constructor for both the adaptation parameters and the initial state.

source

DynamicHMC.adapt_stepsize — Function.

A′ = adapt_stepsize(parameters, A, a)

Update the adaptation A of log stepsize logϵ with average Metropolis acceptance rate a over the whole visited trajectory, using the dual averaging algorithm of Gelman and Hoffman (2014, Algorithm 6). Return the new adaptation.

source

Abstract trajectory interface

In contrast to other reference implementations, the algorithm is implemented in a functional style using immutable values. The intention is to provide more transparency and facilitate fine-grained unit testing.

DynamicHMC.adjacent_tree — Function.

ζ, τ, d, z = adjacent_tree(rng, trajectory, z, depth, fwd)

Traverse the tree of given depth adjacent to point z in trajectory.

fwd specifies the direction, rng is used for random numbers.

Return:

ζ: the proposal from the tree. Only valid when !isdivergent(d) && !isturning(τ), otherwise the value should not be used.

τ: turn statistics. Only valid when !isdivergent(d).

d: divergence statistics, always valid.

z: the point at the end of the tree.

trajectory should support the following interface:

Starting from leaves: ζ, τ, d = leaf(trajectory, z, isinitial)

Moving along the trajectory: z = move(trajectory, z, fwd)

Testing for turning and divergence: isturning(τ), isdivergent(d)

Combination of return values: combine_proposals(ζ₁, ζ₂, bias), combine_turnstats(τ₁, τ₂), and combine_divstats(d₁, d₂)

source

DynamicHMC.Termination — Type.

Reason for terminating a trajectory.

source

DynamicHMC.sample_trajectory — Function.

ζ, d, termination, depth = sample_trajectory(rng, trajectory, z, max_depth)

Sample a trajectory starting at z.

Return:

ζ: proposal from the tree

d: divergence statistics

termination: reason for termination (see Termination)

depth: the depth of the tree that as sampled from. Doubling steps that lead to an invalid tree do not contribute to depth.

See adjacent_tree for the interface that needs to be supported by trajectory.

source

Proposals

DynamicHMC.Proposal — Type.

Proposal that is propagated through by sampling recursively when building the trees.

source

DynamicHMC.combined_logprob_logweight — Function.

logprob, ω = combined_logprob_logweight(ω₁, ω₂, bias)

Given (relative) log probabilities ω₁ and ω₂, return the log probabiliy of drawing a sampel from the second (logprob) and the combined (relative) log probability (ω).

When bias, biases towards the second argument, introducing anti-correlations.

source

DynamicHMC.combine_proposals — Function.

combine_proposals(rng, ζ₁, ζ₂, bias)

Combine proposals from two trajectories, using their weights.

When bias, biases towards the second proposal, introducing anti-correlations.

source

Divergence statistics

DynamicHMC.DivergenceStatistic — Type.

Divergence and acceptance statistics.

Calculated over all visited phase points (not just the tree that is sampled from).

source

DynamicHMC.divergence_statistic — Function.

divergence_statistic()

Empty divergence statistic (for initial node).

source

divergence_statistic(isdivergent, Δ)

Divergence statistic for leaves. Δ is the log density relative to the initial point.

source

DynamicHMC.isdivergent — Function.

isdivergent(x)

Test if divergence statistic x indicates divergence.

source

DynamicHMC.combine_divstats — Function.

combine_divstats(x, y)

Combine divergence statistics from (subtrees) x and y. A divergent subtree make a subtree divergent.

source

Turn analysis

DynamicHMC.TurnStatistic — Type.

Statistics for the identification of turning points. See Betancourt (2017, appendix).

source

DynamicHMC.combine_turnstats — Function.

combine_turnstats(x, y)

Combine turn statistics of two trajectories x and y, which are assume to be adjacent and in that order.

source

DynamicHMC.isturning — Function.

isturning(τ)

Test termination based on turn statistics. Uses the generalized NUTS criterion from Betancourt (2017).

Note that this function should not be called with turn statistics returned by leaf, ie depth > 0 is required.

source

Sampling

DynamicHMC.Trajectory — Type.

Representation of a trajectory, ie a Hamiltonian with a discrete integrator that also checks for divergence.

source

DynamicHMC.leaf — Function.

ζ, τ, d = leaf(trajectory, z, isinitial)

Construct a proposal, turn statistic, and divergence statistic for a single point z in trajectory. When isinitial, z is the initial point in the trajectory.

Return

ζ: the proposal, which should only be used when !isdivergent(d)

τ: the turn statistic, which should only be used when !isdivergent(d)

d: divergence statistic

source

DynamicHMC.move — Function.

move(trajectory, z, fwd)

Return next phase point adjacent to z along trajectory in the direction specified by fwd.

source

DynamicHMC.NUTS_transition — Function.

NUTS_transition(rng, H, q, ϵ, max_depth; args...)

No-U-turn Hamiltonian Monte Carlo transition, using Hamiltonian H, starting at position q, using stepsize ϵ. Builds a doubling dynamic tree of maximum depth max_depth. args are passed to the Trajectory constructor. rng is the random number generator used.

source

Diagnostics

DynamicHMC.NUTS_Statistics — Type.

Storing the output of NUTS_statistics in a structured way, for pretty printing. Currently for internal use.

source

DynamicHMC.ACCEPTANCE_QUANTILES — Constant.

Acceptance quantiles for NUTS_Statistics diagnostic summary.

source

DynamicHMC.explore_local_acceptance_ratios — Function.

explore_local_acceptance_ratios(H, q, ϵs, ps)

Return a matrix of local_acceptance_ratio values for stepsizes ϵs and the given momentums ps. The latter is calculated from random values when an integer is given.

To facilitate plotting, $-∞$ values are replaced by NaN.

source

Utilities and miscellanea

DynamicHMC.rand_bool — Function.

rand_bool(rng, prob)

Random boolean which is true with the given probability prob.

All random numbers in this library are obtained from this function.

source