Low-level building blocks

Notation

Notation follows Betancourt (2017), with some differences.

Instead of energies, negative energies are used in the code.

The following are used consistently for variables:

ℓ: log density we sample from, supports the interface of LogDensityProblems.AbstractLogDensityProblem

κ: distribution/density that corresponds to kinetic energy

H: Hamiltonian

q: position

p: momentum

z: point in phase space (q,p)

ϵ: stepsize

a: acceptance rate

A: acceptance tuning state

ζ: proposal from trajectory (phase point and weight)

τ: turn statistic

d: divergence statistic

π: log density (different from papers)

Δ: logdensity relative to initial point of trajectory

Low-level building blocks

This is documented only for developers. These are not part of the public API, if you are using them you should reconsider or file an issue.

Hamiltonian and leapfrog

Hamiltonian
PhasePoint
phasepoint_in
rand_phasepoint
neg_energy
get_p♯
loggradient
leapfrog

Finding initial stepsize $\epsilon$

Local stepsize tuning.

The local acceptance ratio is technically a probability, but for finding the initial stepsize, it is not capped at $1$.

Also, the values are cached as this is assumed to be moderately expensive to calculate.

DynamicHMC.find_initial_stepsize — Function.

find_initial_stepsize(parameters, A)

Find an initial stepsize that matches the conditions of parameters (see InitialStepsizeSearch).

A is the local acceptance ratio (uncapped). When given a Hamiltonian H and a phasepoint z, it will be calculated using local_acceptance_ratio.

source

DynamicHMC.InitialStepsizeSearch — Type.

struct InitialStepsizeSearch

Parameters for the search algorithm for the initial stepsize.

The algorithm finds an initial stepsize $ϵ$ so that the local acceptance ratio $A(ϵ)$ satisfies

\[a_\text{min} ≤ A(ϵ) ≤ a_\text{max}\]

This is achieved by an initial bracketing, then bisection.

a_min

Lowest local acceptance rate.

a_max

Highest local acceptance rate.

ϵ₀

Initial stepsize.

C

Scale factor for initial bracketing, > 1. Default: 2.0.

maxiter_crossing

Maximum number of iterations for initial bracketing.

maxiter_bisect

Maximum number of iterations for bisection.

Note

Cf. Hoffman and Gelman (2014), which does not ensure bounds for the acceptance ratio, just that it has crossed a threshold. This version seems to work better for some tricky posteriors with high curvature.

source

DynamicHMC.find_crossing_stepsize — Function.

find_crossing_stepsize(parameters, A, ϵ₀)
find_crossing_stepsize(parameters, A, ϵ₀, Aϵ₀)

Find the stepsize for which the local acceptance rate A(ϵ) crosses a.

Return ϵ₀, A(ϵ₀), ϵ₁, A(ϵ₁), whereϵ₀andϵ₁are stepsizes before and after crossingawithA(ϵ)`, respectively.

Assumes that $A(ϵ₀) ∉ (a_\text{min}, a_\text{max})$, where the latter are defined in parameters.

parameters: parameters for the iteration.

A: local acceptance ratio (uncapped), a function of stepsize ϵ

ϵ₀, Aϵ₀: initial value of ϵ, and A(ϵ₀)

source

DynamicHMC.bisect_stepsize — Function.

bisect_stepsize(parameters, A, ϵ₀, ϵ₁)
bisect_stepsize(parameters, A, ϵ₀, ϵ₁, Aϵ₀)
bisect_stepsize(parameters, A, ϵ₀, ϵ₁, Aϵ₀, Aϵ₁)

Return the desired stepsize ϵ by bisection.

parameters: algorithm parameters, see InitialStepsizeSearch

A: local acceptance ratio (uncapped), a function of stepsize ϵ

ϵ₀, ϵ₁, Aϵ₀, Aϵ₁: stepsizes and acceptance rates (latter optional).

This function assumes that $ϵ₀ < ϵ₁$, the stepsize is not yet acceptable, and the cached A values have the correct ordering.

source

DynamicHMC.local_acceptance_ratio — Function.

local_acceptance_ratio(H, z)

Return a function of the stepsize ($ϵ$) that calculates the local acceptance ratio for a single leapfrog step around z along the Hamiltonian H. Formally, let

A(ϵ) = exp(logdensity(H, leapfrog(H, z, ϵ)) - logdensity(H, z))

Note that the ratio is not capped by 1, so it is not a valid probability per se.

source

Dual averaging

DualAveragingParameters
DualAveragingAdaptation
get_ϵ
adapting_ϵ
adapt_stepsize

Abstract trajectory interface

In contrast to other reference implementations, the algorithm is implemented in a functional style using immutable values. The intention is to provide more transparency and facilitate fine-grained unit testing.

adjacent_tree
Termination
sample_trajectory

Proposals

Proposal
combined_logprob_logweight
combine_proposals

Divergence statistics

DivergenceStatistic
divergence_statistic
isdivergent
combine_divstats

Turn analysis

TurnStatistic
combine_turnstats
isturning

Sampling

Trajectory
leaf
move
NUTS_transition

Tuning

DynamicHMC.AbstractTuner

Diagnostics

NUTS_Statistics
ACCEPTANCE_QUANTILES
explore_local_acceptance_ratios

Reporting information during runs

Samplers take an AbstractReport argument, which is then used for reporting. The interface is as follows.

DynamicHMC.AbstractReport
DynamicHMC.report!
DynamicHMC.start_progress!
DynamicHMC.end_progress!

The default is

ReportIO

Reporting information can be suppressed with

ReportSilent

Other interfaces should define similar types.

Utilities and miscellanea

DynamicHMC.rand_bool — Function.

rand_bool(rng, prob)

Random boolean which is true with the given probability prob.

source